requirements.txt
pandas
scikit-learn
streamlit
joblib
gunicorn
flask
numpy
requests
instaloader
flask-login
flask-bcrypt
flask-limiter
python-dotenv